+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Implemented FEP-ef61 resolver endpoint.
+- Use actor ID as handle if acct URI is not known.
 
 ### Changed